quick bug fixes in front end code

+++ metapredict/tests/test_cython_v_python.py
@@ -1,10 +1,8 @@
 # Import package, test suite, and other packages as needed
 import metapredict
 from metapredict import meta
-import metapredict.backend.uniprot_predictions
 
 import numpy as np
-
 
 from metapredict.backend.cython.domain_definition import build_domains_from_values 
 from metapredict.backend.domain_definition import __build_domains_from_values as  domain_definition
@@ -25,7 +23,7 @@ def test_domain_decomposition():
     for _ in range(50):
         s = build_seq()
     
-        disorder = meta.predict_disorder(s, return_numpy=True)
+        disorder = meta.predict_disorder(s, return_numpy=True).astype(np.double)
     
         for thresh in [0, 0.1, 0.2, 0.5, 0.9, 1.0]:
             cyth = build_domains_from_values(disorder, thresh)
